backup/UseCursorEditMOdy/mysql_manager/requirements.txt
PyQt6==6.6.1
mysql-connector-python==8.3.0
pandas==2.2.0 